+++ system/core/Handlers.class.php
@@ -8,7 +8,7 @@
     activates them depending on the requested URI.
 
     @package urlaube\urlaube
-    @version 0.1a11
+    @version 0.1a12
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */