+++ system/core/Main.class.php
@@ -7,7 +7,7 @@
     handles the actual workflow of the urlau.be CMS.
 
     @package urlaube\urlaube
-    @version 0.1a11
+    @version 0.1a12
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */