+++ system/handlers/fixurl/handler.php
@@ -7,7 +7,7 @@
     handler is meant to improve incorrectly written URLs.
 
     @package urlaube\urlaube
-    @version 0.1a11
+    @version 0.1a12
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */