+++ system/plugins/hide/plugin.php
@@ -7,7 +7,7 @@
     introduces the hide feature to remove content from certain handler results.
 
     @package urlaube\urlaube
-    @version 0.1a11
+    @version 0.1a12
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */