+++ system/plugins/markdown/plugin.php
@@ -7,7 +7,7 @@
     converts markdown-encoded content to HTML-encoded content.
 
     @package urlaube\urlaube
-    @version 0.1a11
+    @version 0.1a12
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */